Updated dependencies; Made compiler warnings stricter

+++ src/encryptor.cpp
@@ -940,7 +940,7 @@ EncryptResult Encryptor::EncryptStream(
                 // Pad with value 16 - read_length octets
                 std::fill(plaintext.data() + read_length,
                           plaintext.data() + plaintext.size(),
-                          16 - read_length);
+                          static_cast<std::uint8_t>(16 - read_length));
             }
 
             // Encrypt the block
